⚡ chore: move config to /src
