--- src/app/index.ts
@@ -1,1 +1,0 @@
-export { config } from './config';
