+++ src/config/sections/music.ts
