+++ src/config/sections/pacman.ts
