+++ src/config/sections/snake.ts
