+++ src/config/sections/techs.ts
